Added Spring data JPA support(Repositoris, services), big changes to domain objects

+++ src/main/java/nc/apps/dao/BookBaseModelDAOImpl.java
@@ -35,7 +35,7 @@ public class BookBaseModelDAOImpl implements BookBaseModelDAO {
             PreparedStatement statement = con.prepareStatement(SQL_GET_ALL)){
             ResultSet resultSet = statement.executeQuery();
             while(resultSet.next()){
-                bookBaseModels.add(new BookBaseModel(resultSet.getLong("id"),
+                bookBaseModels.add(new BookBaseModel(resultSet.getInt("id"),
                         resultSet.getString("title")));
             }
         } catch (SQLException e) {
